image-generator/image-generator/requirements.txt
Bootstrap_Flask
Flask_CKEditor
flask-bootstrap
Flask_Login
Flask-Gravatar
Flask_WTF
WTForms
Werkzeug
Flask
flask_sqlalchemy
flask-reuploaded
SQLAlchemy
gunicorn
psycopg2-binary
numpy
Pillow
webcolors
